requirements.txt
discord.py==2.0.1
python_Levenshtein==0.20.8
pymongo==4.3.2
